Fix the blue for the icon's hover state, move color filter functions into utils, change colorHoverLink to a constant
