+++ app/assets/src/components/visualizations/heatmap/Heatmap.js
@@ -11,6 +11,9 @@ import cx from "classnames";
 import symlog from "../../utils/d3/scales/symlog.js";
 import cs from "./heatmap.scss";
 import { CategoricalColormap } from "../../utils/colormaps/CategoricalColormap.js";
+import addSvgColorFilter from "../../utils/d3/svg.js";
+// used for filter to make plus icon blue
+const COLOR_HOVER_LINK = "#3867fa";
 
 // TODO(tcarvalho): temporary hack to send elements to the back.
 // Remove once code is ported to d3 v4, which contains this function.
@@ -55,7 +58,6 @@ export default class Heatmap {
         maxColumnClusterHeight: 100,
         spacing: 10,
         metadataAddLinkHeight: 14,
-        colorHoverLink: "#2b52cd", // used for filter to make icon blue
         transitionDuration: 200,
         nullValue: 0,
         columnMetadata: [],
@@ -269,37 +271,8 @@ export default class Heatmap {
     );
 
     const defs = this.svg.append("defs");
-
-    // Creates a color matrix that can recolor a black icon via a filter. See:
-    // https://semisignal.com/using-fecolormatrix-to-dynamically-recolor-icons-part-1-single-color-icons/
-    const generateColorMatrix = function(rgb) {
-      let rScaled = rgb[0] / 255.0;
-      let gScaled = rgb[1] / 255.0;
-      let bScaled = rgb[2] / 255.0;
-      return `0 0 0 0 ${rScaled}
-              0 0 0 0 ${gScaled}
-              0 0 0 0 ${bScaled}
-              0 0 0 1 0`;
-    };
-
-    // Convert hex string to rgb array
-    const hexToRgb = function(hex) {
-      var r = parseInt(hex.slice(1, 3), 16),
-        g = parseInt(hex.slice(3, 5), 16),
-        b = parseInt(hex.slice(5, 7), 16);
-      return [r, g, b];
-    };
-
     // Create a blue color filter to match $primary-light.
-    defs
-      .append("filter")
-      .attr("id", "blue")
-      .append("feColorMatrix")
-      .attr("type", "matrix")
-      .attr(
-        "values",
-        generateColorMatrix(hexToRgb(this.options.colorHoverLink))
-      );
+    addSvgColorFilter(defs, "blue", COLOR_HOVER_LINK);
 
     this.g = this.svg.append("g");
     this.gRowLabels = this.g.append("g").attr("class", cs.rowLabels);
